# perf: make add_mod faster using overflow check