--- a/src/modular.rs
+++ b/src/modular.rs
@@ -7,6 +7,21 @@ use crate::{algorithms, Uint};
 // See also <https://static1.squarespace.com/static/61f7cacf2d7af938cad5b81c/t/62deb4e0c434f7134c2730ee/1658762465114/modular_multiplication.pdf>
 // FEATURE: Modular wrapper class, like Wrapping.
 
+/// `let $id = &mut [0u64; nlimbs(2 * BITS)][..]`
+macro_rules! let_double_bits {
+    ($id:ident) => {
+        // This array casting is a workaround for `generic_const_exprs` not being
+        // stable.
+        let mut double = [[0u64; 2]; LIMBS];
+        let double_len = crate::nlimbs(2 * BITS);
+        debug_assert!(2 * LIMBS >= double_len);
+        // SAFETY: `[[u64; 2]; LIMBS] == [u64; 2 * LIMBS] >= [u64; nlimbs(2 * BITS)]`.
+        let $id = unsafe {
+            core::slice::from_raw_parts_mut(double.as_mut_ptr().cast::<u64>(), double_len)
+        };
+    };
+}
+
 impl<const BITS: usize, const LIMBS: usize> Uint<BITS, LIMBS> {
     /// ⚠️ Compute $\mod{\mathtt{self}}_{\mathtt{modulus}}$.
     ///
@@ -48,32 +63,12 @@ impl<const BITS: usize, const LIMBS: usize> Uint<BITS, LIMBS> {
         // do overflowing add, then check if we should divrem
         let (result, overflow) = self.overflowing_add(rhs);
         if overflow {
-            // Allocate at least `nlimbs(2 * BITS)` limbs to store the numerator. This array
-            // casting is a workaround for `generic_const_exprs` not being stable.
-            let mut numerator = [[0u64; 2]; LIMBS];
-            let numerator_len = crate::nlimbs(2 * BITS);
-            debug_assert!(2 * LIMBS >= numerator_len);
-            // SAFETY: `[[u64; 2]; LIMBS] == [u64; 2 * LIMBS] >= [u64; nlimbs(2 * BITS)]`.
-            let numerator = unsafe {
-                core::slice::from_raw_parts_mut(numerator.as_mut_ptr().cast::<u64>(), numerator_len)
-            };
-
-            // Copy the result limbs into the numerator.
-            // SAFETY:
-            // * `result` is valid for the length of `result.limbs`
-            // * `numerator` is valid for the length of `result.limbs`
-            // * `result` and `numerator` do not overlap
-            unsafe {
-                core::ptr::copy_nonoverlapping(
-                    result.as_limbs().as_ptr(),
-                    numerator.as_mut_ptr(),
-                    result.limbs.len(),
-                );
-            };
-
-            // Now we have to fill in the carry bit before reducing.
-            let (limbs, bits) = (BITS / 64, BITS % 64);
-            numerator[limbs] |= 1 << bits;
+            // Add carry bit to the result in an extra limb.
+            let_double_bits!(numerator);
+            let numerator = &mut numerator[..=LIMBS];
+            numerator[..LIMBS].copy_from_slice(result.as_limbs());
+            let (limb, bit) = (BITS / 64, BITS % 64);
+            numerator[limb] |= 1 << bit;
 
             // Compute modulus using `div_rem`.
             // This stores the remainder in the divisor, `modulus`.
@@ -98,17 +93,8 @@ impl<const BITS: usize, const LIMBS: usize> Uint<BITS, LIMBS> {
             return Self::ZERO;
         }
 
-        // Allocate at least `nlimbs(2 * BITS)` limbs to store the product. This array
-        // casting is a workaround for `generic_const_exprs` not being stable.
-        let mut product = [[0u64; 2]; LIMBS];
-        let product_len = crate::nlimbs(2 * BITS);
-        debug_assert!(2 * LIMBS >= product_len);
-        // SAFETY: `[[u64; 2]; LIMBS] == [u64; 2 * LIMBS] >= [u64; nlimbs(2 * BITS)]`.
-        let product = unsafe {
-            core::slice::from_raw_parts_mut(product.as_mut_ptr().cast::<u64>(), product_len)
-        };
-
         // Compute full product.
+        let_double_bits!(product);
         let overflow = algorithms::addmul(product, self.as_limbs(), rhs.as_limbs());
         debug_assert!(!overflow);