@DaniPopes 2025-05-28
Ready for review, +10-15% general case