@DaniPopes 2025-05-28
Manual benchmark on x86_64 zen 5


```
add_mod/64              time:   [1.8367 ns 1.8400 ns 1.8438 ns]
                        change: [-76.719% -76.655% -76.584%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  4 (4.00%) high mild
  3 (3.00%) high severe

add_mod/128             time:   [27.202 ns 27.211 ns 27.221 ns]
                        change: [-4.9204% -4.8461% -4.7712%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  1 (1.00%) high mild
  3 (3.00%) high severe

add_mod/192             time:   [35.103 ns 35.134 ns 35.166 ns]
                        change: [-19.611% -19.450% -19.311%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  3 (3.00%) high mild
  1 (1.00%) high severe

add_mod/256             time:   [35.361 ns 35.400 ns 35.444 ns]
                        change: [-24.808% -24.602% -24.396%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  7 (7.00%) high mild
  1 (1.00%) high severe

add_mod/384             time:   [47.998 ns 48.327 ns 48.803 ns]
                        change: [-17.419% -17.115% -16.719%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  1 (1.00%) high mild
  3 (3.00%) high severe

add_mod/512             time:   [42.238 ns 42.304 ns 42.387 ns]
                        change: [-26.113% -25.905% -25.675%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  2 (2.00%) low mild
  2 (2.00%) high mild
  2 (2.00%) high severe

add_mod/4096            time:   [158.95 ns 163.22 ns 167.96 ns]
                        change: [-17.772% -16.041% -14.064%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) high mild
  2 (2.00%) high severe

```